# Improve identity projection when the selectedPositions is a list