This PR optimizes identity projection with two ways:

1. Create DictionaryBlock when the selectedPositions is a list;
2. Use a faster and more memory efficient block size calculation.

BenchmarkPageProcessor shows 21x improvement in throughput, 46x reduction in gc.alloc.rate.norm, and 9x reduction in
GC count and time:

Before

```
    Benchmark                                                                    Mode  Cnt       Score      Error   Units
    BenchmarkPageProcessor.identityProjection                                   thrpt   50  125261.816 ± 5342.200   ops/s
    BenchmarkPageProcessor.identityProjection:·gc.alloc.rate                    thrpt   50    2439.004 ±  103.990  MB/sec
    BenchmarkPageProcessor.identityProjection:·gc.alloc.rate.norm               thrpt   50   40832.029 ±    0.033    B/op
    BenchmarkPageProcessor.identityProjection:·gc.churn.PS_Eden_Space           thrpt   50    2461.428 ±  364.315  MB/sec
    BenchmarkPageProcessor.identityProjection:·gc.churn.PS_Eden_Space.norm      thrpt   50   41439.660 ± 6521.978    B/op
    BenchmarkPageProcessor.identityProjection:·gc.churn.PS_Survivor_Space       thrpt   50       0.016 ±    0.016  MB/sec
    BenchmarkPageProcessor.identityProjection:·gc.churn.PS_Survivor_Space.norm  thrpt   50       0.257 ±    0.267    B/op
    BenchmarkPageProcessor.identityProjection:·gc.count                         thrpt   50      46.000             counts
    BenchmarkPageProcessor.identityProjection:·gc.time                          thrpt   50     181.000                 ms
```

After

```
    Benchmark                                                                    Mode  Cnt        Score        Error   Units
    BenchmarkPageProcessor.identityProjection                                   thrpt   10  2687127.250 ± 337956.918   ops/s
    BenchmarkPageProcessor.identityProjection:·gc.alloc.rate                    thrpt   10     1118.285 ±    141.900  MB/sec
    BenchmarkPageProcessor.identityProjection:·gc.alloc.rate.norm               thrpt   10      872.001 ±      0.005    B/op
    BenchmarkPageProcessor.identityProjection:·gc.churn.PS_Eden_Space           thrpt   10     1050.013 ±   1683.604  MB/sec
    BenchmarkPageProcessor.identityProjection:·gc.churn.PS_Eden_Space.norm      thrpt   10      827.947 ±   1329.288    B/op
    BenchmarkPageProcessor.identityProjection:·gc.churn.PS_Survivor_Space       thrpt   10        0.012 ±      0.040  MB/sec
    BenchmarkPageProcessor.identityProjection:·gc.churn.PS_Survivor_Space.norm  thrpt   10        0.011 ±      0.034    B/op
    BenchmarkPageProcessor.identityProjection:·gc.count                         thrpt   10        5.000               counts
    BenchmarkPageProcessor.identityProjection:·gc.time                          thrpt   10       21.000                   ms
```


```
== NO RELEASE NOTE ==
```