@yingsu00 2020-10-09
@mbasmanova just FYI I'm fixing the assertBlockSize() test to cover the case mistakenly used `positionCount` instead of `length` in Map and RowBlocks. Will leave a message once it's done.