@yingsu00 2020-10-13
@mbasmanova Masha, I have updated commit "Add getApproximateRegionLogicalSizeInBytes() to Block " and it's ready for review again. Thank you very much for your careful reading!